增加功能：支持声明式二进制协议

+++ src/main/java/org/jetlinks/protocol/official/binary2/FieldWriter.java
@@ -4,6 +4,6 @@ import io.netty.buffer.ByteBuf;
 
 public interface FieldWriter {
 
-    int write(FieldInstance instance, ByteBuf outputBuf);
+    short write(FieldInstance instance, ByteBuf outputBuf);
 
 }
